+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - It is now possible to register multiple exception handleres to handle different exceptions in app.
 - Major API and command execution changes: a) added `ICliModeSwitcher`, `ICliMode`, and more; b) removed `InteractiveCliApplication`.
 - Replaced `IsInteractiveModeOnly` with `SupportedModes`.
+- Added support for options with no name by automatic conversion of property names. 
 
 ### v2.1.1 (18-Oct-2020)